fw: ForwarderPurifProc and ForwarderSwapProc
