+++ mqns/network/fw/__init__.py
@@ -1,8 +1,8 @@
-from mqns.network.fw.classic import fw_control_cmd_handler, fw_signaling_cmd_handler
 from mqns.network.fw.controller import RoutingController
 from mqns.network.fw.cutoff import CutoffScheme, CutoffSchemeWaitTime, CutoffSchemeWaitTimeCounters
 from mqns.network.fw.fib import Fib, FibEntry
 from mqns.network.fw.forwarder import Forwarder, ForwarderCounters
+from mqns.network.fw.fw_classic import fw_control_cmd_handler, fw_signaling_cmd_handler
 from mqns.network.fw.message import MultiplexingVector, SwapSequence
 from mqns.network.fw.mux import MuxScheme
 from mqns.network.fw.mux_buffer_space import MuxSchemeBufferSpace
